seatalk-intraday-bot/requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
pillow
python-dotenv
schedulej